Revert this collectstatic if this doesn't work! workaround for bootstrap obsessed command
Having tried every solution under the sun, collectstatic continues to assign a bunch of requirements using Bootstrap, which I'm not using on this project. But without it, it can't work properly as it's a part of the DRF. So I'm trying an ugly workaround of installing bootstrap with no intention of using it, along with creating a load of empty files that the collectstatic command was crying after. Here goes nothing, let's hope this solves the deployment issues where the admin panel had no CSS!